requirements.txt
Django==4.2.7
mysqlclient==2.1.1
Pillow==10.0.1
opencv-python==4.8.1.78
matplotlib==3.7.2
pandas==2.1.1
reportlab==4.0.5
openpyxl==3.1.2
celery==5.3.4
redis==5.0.1
django-crispy-forms==2.0
crispy-bootstrap5==0.7
python-decouple==3.8
weasyprint==61.0
numpy==1.24.3

